requirements.txt
bson==0.5.10
click==7.1.2
dnspython==1.16.0
Flask==1.1.2
Flask-Cors==3.0.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pymongo==3.11.2
python-dateutil==2.8.1
python-dotenv==0.15.0
six==1.15.0
Werkzeug==1.0.1
